requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
Django==3.0.3
django-markdown-deux==1.0.5
django-pagedown==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
markdown2==2.3.8
mccabe==0.6.1
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
